mainファイル名・DB名変更。
